% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blog_name.R
\name{blog_name}
\alias{blog_name}
\title{Code a name}
\usage{
blog_name(name, gap = 1)
}
\arguments{
\item{name}{a character}

\item{gap}{code gap}
}
\value{
a character
}
\description{
Code a name
}
\examples{
blog_name("gdkkn")
blog_name("eg", gap = 2)

}
