% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_tag.R
\name{clip_convert_commas}
\alias{clip_convert_commas}
\title{Converts Chinese inverted commas to right angle inverted commas}
\usage{
clip_convert_commas(reverse = FALSE)
}
\arguments{
\item{reverse}{Logical. The default is \code{FALSE}, which converts inverted commas to right-angle inverted commas; if \code{TRUE}, it converts right-angle inverted commas to inverted commas.}
}
\value{
a piece of text to your Clipboard.
}
\description{
Converts Chinese inverted commas to right angle inverted commas
}
\examples{
\dontrun{
Sys.setenv(CLIPR_ALLOW=TRUE)
clipr::write_clip("Sentences-containing-Chinese-inverted-commas.", allow_non_interactive = TRUE)
clip_convert_commas()
clipr::read_clip()
}
}
