% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_tag.R
\name{clip_remove_spaces}
\alias{clip_remove_spaces}
\title{Remove spaces between Chinese characters and between Chinese characters and punctuation}
\usage{
clip_remove_spaces()
}
\value{
a piece of text to you Clipboard.
}
\description{
Remove spaces between Chinese characters and between Chinese characters and punctuation
}
\examples{
\dontrun{
Sys.setenv(CLIPR_ALLOW=TRUE)
clipr::write_clip("\\\\u4f60 \\\\u597d\\\\u554a \\\\uff01", allow_non_interactive = TRUE)
clip_remove_spaces()
clipr::read_clip()
}
}
