% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_tag.R
\name{clip_remove_tag}
\alias{clip_remove_tag}
\title{Remove links and tags}
\usage{
clip_remove_tag()
}
\value{
a piece of text to your Clipboard.
}
\description{
Read from the Clipboard, remove all \code{[[}, \verb{]]}
and \verb{#tag}, which is widely used in note taking tools like
\emph{Roam Research} and write to Clipboard again.
}
\examples{
\dontrun{
Sys.setenv(CLIPR_ALLOW=TRUE)
clipr::write_clip("A [[nice]] day. #happy", allow_non_interactive = TRUE)
clip_remove_tag()
clipr::read_clip()
}
}
