% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_score.R
\name{cut_value}
\alias{cut_value}
\title{Calculate the value of x based on the cut points and corresponding segment values}
\usage{
cut_value(x, breaks, values, right = TRUE)
}
\arguments{
\item{x}{A numeric value.}

\item{breaks}{A numeric vector indicating the cut points, left open and right closed.}

\item{values}{A vector indicating the return values of each segment, its length must be equal to \code{breaks} + 1.}

\item{right}{Logical, indicating if the intervals should be closed on the right (and open on the left) or vice versa.}
}
\value{
A value from \code{values}.
}
\description{
Calculate the value of x based on the cut points and corresponding segment values
}
\examples{
cut_value(2, 1, c(1, -1))
cut_value(2, 1, LETTERS[1:2])
cut_value(2, c(1, 3), c(1, 3, 5))
}
