% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_tag.R
\name{insert_spaces}
\alias{insert_spaces}
\title{Insert spaces between numbers and Chinese}
\usage{
insert_spaces()
}
\value{
a piece of text to you Clipboard.
}
\description{
Insert spaces between numbers and Chinese
}
