% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kindle_clean.R
\name{kindle_clean}
\alias{kindle_clean}
\title{Clean Kindle file}
\usage{
kindle_clean(
  data,
  distinct = TRUE,
  drop_na = TRUE,
  new_line = TRUE,
  title = NULL
)
}
\arguments{
\item{data}{The data read from \code{\link[=kindle_read]{kindle_read()}}.}

\item{distinct}{Logical. If \code{TRUE}, the last record with the same \code{begin}
value will be retained; if \code{FALSE}, no records will be deleted.}

\item{drop_na}{Logical. If \code{TRUE}, drop rows where \code{text} column is \code{NA}; if \code{FALSE},
no records will be deleted.}

\item{new_line}{Logical. If \code{TRUE}, \verb{ } will be replaced by \verb{\\n} which can create
a new line.}

\item{title}{String. Provide the \code{title} of the \code{data} you want to write out,
with support for regular expressions. If \code{NULL}, all of records will be wrote out.}
}
\value{
A tidy data.
}
\description{
Clean Kindle file
}
\examples{
\dontrun{
text <- kindle_read("tests/testthat/My-Clippings.txt")
kindle_clean(text)
}

}
