% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kindle_read.R
\name{kindle_read}
\alias{kindle_read}
\title{Read \verb{My Clippings.txt} from Kindle and clean it into tidy data}
\usage{
kindle_read(file)
}
\arguments{
\item{file}{The path to \verb{My Clippings.txt}, generally in \verb{Kindle/Documents/}}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
Read \verb{My Clippings.txt} from Kindle and clean it into tidy data
}
\examples{
\dontrun{
kindle_read("tests/testthat/My-Clippings.txt")
}
}
