% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kindle_write.R
\name{kindle_write_md}
\alias{kindle_write_md}
\title{Export to Markdown}
\usage{
kindle_write_md(
  data,
  file,
  title = NULL,
  distinct = TRUE,
  drop_na = TRUE,
  new_line = TRUE,
  time = TRUE
)
}
\arguments{
\item{data}{The data read from \code{\link[=kindle_read]{kindle_read()}}.}

\item{file}{The file name to write to, book title is recommended.}

\item{title}{String. Provide the \code{title} of the \code{data} you want to write out,
with support for regular expressions.}

\item{distinct}{Logical. If \code{TRUE}, the last record with the same \code{begin}
value will be retained; if \code{FALSE}, no records will be deleted.}

\item{drop_na}{Logical. If \code{TRUE}, drop rows where \code{text} column is \code{NA}; if \code{FALSE},
no records will be deleted.}

\item{new_line}{Logical. If \code{TRUE}, \verb{ } will be replaced by \verb{\\n} which can create
a new line.}

\item{time}{Logical. If \code{TRUE}, the \code{datetime} will be add to the end of \code{text}.}
}
\value{
A Markdown file.
}
\description{
Write out partial data to a single Markdown file.
}
\examples{
\dontrun{
kindle_write_md(text, "mdtest.md", title = "book-title")
}
}
