% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kindle_write.R
\name{kindle_write_xlsx}
\alias{kindle_write_xlsx}
\title{Export to xlsx}
\usage{
kindle_write_xlsx(
  data,
  file,
  columns = NULL,
  distinct = TRUE,
  drop_na = TRUE,
  new_line = TRUE,
  title = NULL
)
}
\arguments{
\item{data}{The data read from \code{\link[=kindle_read]{kindle_read()}}.}

\item{file}{The file name to write to.}

\item{columns}{A character vector. Select and reorder variables in a data frame. Choose from the following values:
\itemize{
\item \code{id}: The row number.
\item \code{title}: Name of the book.
\item \code{page}: The number of page of the excerpt in the book.
\item \code{begin}: The starting position of the excerpt in the book.
\item \code{end}: The ending position of the excerpt in the book.
\item \code{datetime}: The time when the excerpt is took down.
\item \code{text}: Excerpted content.
}}

\item{distinct}{Logical. If \code{TRUE}, the last record with the same \code{begin}
value will be retained; if \code{FALSE}, no records will be deleted.}

\item{drop_na}{Logical. If \code{TRUE}, drop rows where \code{text} column is \code{NA}; if \code{FALSE},
no records will be deleted.}

\item{new_line}{Logical. If \code{TRUE}, \verb{ } will be replaced by \verb{\\n} which can create
a new line.}

\item{title}{String. Provide the \code{title} of the \code{data} you want to write out,
with support for regular expressions. If \code{NULL}, all of records will be wrote out.}
}
\value{
A excel file.
}
\description{
Export to xlsx
}
\examples{
\dontrun{
text <- kindle_read("tests/testthat/My-Clippings.txt")
kindle_write_xlsx(text, "tests/testthat/kindle_write.xlsx")
}

}
