% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life_answers.R
\name{life_answers}
\alias{life_answers}
\title{The answers of your life}
\usage{
life_answers(n = 1)
}
\arguments{
\item{n}{The number of answers returned.}
}
\value{
Some text
}
\description{
The answers of your life
}
\examples{
\dontrun{
life_answer()
life)answer(n = 2)
}
}
