% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentile.R
\name{percentile}
\alias{percentile}
\title{Calculate the percentile of value x in a set of numbers}
\usage{
percentile(x, vector, digits = 4, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric value.}

\item{vector}{A set of numeric values.}

\item{digits}{Integer indicating the number of decimal places (round) or significant digits (signif) to be used.}

\item{na.rm}{Logical. Should missing values (including NaN) be removed?}
}
\value{
A numeric value.
}
\description{
Calculate the percentile of value x in a set of numbers
}
\examples{
percentile(2, c(1, 5))
percentile(2, c(1, 3, 5), digits = 2)
}
