% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{progress}
\alias{progress}
\title{Learning Progress}
\usage{
progress(now, total)
}
\arguments{
\item{now}{The current progress of learning.}

\item{total}{The total amount of learning materials.}
}
\value{
A string indicating the learning progress.
}
\description{
Calculate and display the learning progress.
}
\examples{
progress(50, 100)

}
