% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_es.R
\name{read_es_query}
\alias{read_es_query}
\title{Read Data from Elasticsearch Query JSON}
\usage{
read_es_query(json_file)
}
\arguments{
\item{json_file}{The path to the JSON file containing the Elasticsearch query results.}
}
\value{
Returns a tibble containing the data extracted from the query results.
}
\description{
This function is used to extract data from a JSON file containing the results of an Elasticsearch query. It converts the results into a tibble, then flattens the '_source' column and removes the '_index', '_id', and '_score' columns.
}
\examples{
\dontrun{
# Read Elasticsearch query results from a JSON file
query_result <- read_es_query("path/to/query_result.json")

# View the first few rows of data
head(query_result)
}

}
