% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_message.R
\name{send_message_ding}
\alias{send_message_ding}
\title{Send a message to DingTalk robot}
\usage{
send_message_ding(
  message,
  include_timestamp = FALSE,
  at_mobiles = NULL,
  at_all = FALSE
)
}
\arguments{
\item{message}{The message content to send.}

\item{include_timestamp}{Whether to include the current timestamp in the message. Default is \code{FALSE}.}

\item{at_mobiles}{A character vector of mobile numbers to mention in the message. Default is \code{NULL}.}

\item{at_all}{Whether to mention all users in the message. Default is \code{FALSE}.}
}
\description{
This function sends a message to a DingTalk robot using the provided webhook URL.
}
\examples{
\dontrun{
options(webhook_ding = "https://oapi.dingtalk.com/robot/send?access_token=YOUR_ACCESS_TOKEN")
send_message_ding("Hello, World!")
send_message_ding("Code execution completed.", include_timestamp = TRUE,
 at_mobiles = c("1234567890", "0987654321"))
send_message_ding("Code execution completed.", include_timestamp = TRUE, at_all = TRUE)
}

}
\seealso{
\url{https://open.dingtalk.com/document/robots/custom-robot-access}
}
\keyword{dingtalk}
\keyword{message}
