% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_message.R
\name{send_message_wx}
\alias{send_message_wx}
\title{Send a message to Enterprise WeChat}
\usage{
send_message_wx(
  message,
  include_timestamp = FALSE,
  at_mobiles = NULL,
  at_all = FALSE
)
}
\arguments{
\item{message}{The message content to be sent.}

\item{include_timestamp}{Logical indicating whether to include the current timestamp in the message. Default is \code{FALSE}.}

\item{at_mobiles}{A character vector of mobile numbers to be mentioned in the message. Default is \code{NULL}.}

\item{at_all}{Logical indicating whether to mention all members. Default is \code{FALSE}.}
}
\value{
None
}
\description{
This function sends a message to an Enterprise WeChat robot using the provided webhook URL.
}
\examples{
\dontrun{
options(webhook_wx = "https://qyapi.weixin.qq.com/robot/send?access_token=YOUR_ACCESS_TOKEN")
send_message_wx("Code execution completed.", include_timestamp = TRUE, at_mobiles = "1234567890")
send_message_wx("Code execution completed.", include_timestamp = TRUE, at_all = TRUE)
}

}
