% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_cal.R
\name{state_info}
\alias{state_info}
\title{Calculate Financial Statement Date Range and Type}
\usage{
state_info(begin_date, end_date)
}
\arguments{
\item{begin_date}{A string representing the start date of the financial statement period. The format should be in 'YYYY-MM-DD'.}

\item{end_date}{A string representing the end date of the financial statement period. The format should be in 'YYYY-MM-DD'.}
}
\value{
A string which contains the date range and type of the financial statement.
}
\description{
Calculate Financial Statement Date Range and Type
}
\examples{
state_info("2023-01-01", "2023-12-31")
}
