% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_cal.R
\name{state_type}
\alias{state_type}
\title{Calculate Financial Statement Type}
\usage{
state_type(begin_date, end_date)
}
\arguments{
\item{begin_date}{A string representing the start date of the financial statement period. The format should be in 'YYYY-MM-DD'.}

\item{end_date}{A string representing the end date of the financial statement period. The format should be in 'YYYY-MM-DD'.}
}
\value{
A string which contains the type of the financial statement.
}
\description{
Calculate Financial Statement Type
}
\examples{
state_type("2023-01-01", "2023-12-31")
}
